+++ FInalProj/bin/main/templates/views/common/header.html
@@ -33,7 +33,7 @@
     </header>
     <section id="subMenu">
         <div>
-            <div><a href="noticeMain.bo">NOTICE</a></div>
+            <div><a href="noticeList.bo">NOTICE</a></div>
             <div><a href="fineNewsMain.bo">선한 뉴스</a></div>
             <div><a href="finePeopleMain.bo">선뜻한 사람</a></div>
             <div><a href="fruitMain.bo">결실</a></div>
@@ -45,9 +45,9 @@
             <div><a href="supportMain.su">모금함</a></div>
         </div>
         <div>
-            <div><a href="talkMain.bo">TALK</a></div>
+            <div><a href="commList.bo">TALK</a></div>
             <div><a href="faqMain.bo">자주 묻는 질문</a></div>
-            <div><a href="qnaMain.bo">Q&A</a></div>
+            <div><a href="qaList.bo">Q&A</a></div>
         </div>
         <div>
             <!-- 비 로그인 -->
