--- FInalProj/src/main/java/com/fin/proj/community/model/dao/CommunityDAO.java
@@ -1,5 +1,0 @@
-package com.fin.proj.community.model.dao;
-
-public interface CommunityDAO {
-
-}
