--- FInalProj/src/main/java/com/fin/proj/community/model/service/CommunityService.java
@@ -1,5 +1,0 @@
-package com.fin.proj.community.model.service;
-
-public class CommunityService {
-
-}
