--- FInalProj/src/main/java/com/fin/proj/community/model/service/CommunityServiceImpl.java
@@ -1,8 +1,0 @@
-package com.fin.proj.community.model.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface CommunityServiceImpl {
-
-}
