+++ FInalProj/src/main/resources/templates/views/board/commDetail.html
@@ -93,9 +93,9 @@
         
          <div class="actionArea">
             <div class="writeArea">
-                <button class="writeFaqBtn" onclick="location.href='${ contextPath }/commList.co'">목록으로</button>
-                <button class="writeFaqBtn" onclick="location.href='${ contextPath }/editComm.co'">수정하기</button>
-                <button class="writeFaqBtn" onclick="location.href='${ contextPath }/commList.co'">삭제하기</button>
+                <button class="writeFaqBtn" onclick="location.href='commList.bo'">목록으로</button>
+                <button class="writeFaqBtn" onclick="location.href='editComm.bo'">수정하기</button>
+                <button class="writeFaqBtn" onclick="location.href='commList.bo'">삭제하기</button>
             </div>
 			</div>
     </section>
