+++ FInalProj/src/main/resources/templates/views/board/noticeList.html
@@ -49,7 +49,7 @@
 
         <div class="actionArea">
             <div class="writeArea">
-                <button class="writeFaqBtn" onclick="location.href='writeNotice.co'">작성하기</button>
+                <button class="writeFaqBtn" onclick="location.href='writeNotice.bo'">작성하기</button>
             </div>
             <div class="searchArea">
                 <input class="searchInput" type="text" placeholder="검색"/>
