+++ FInalProj/src/main/resources/templates/views/board/qaDetail.html
@@ -49,10 +49,10 @@
         </table>
          <div class="actionArea">
             <div class="writeArea">
-                <button class="writeFaqBtn" onclick="location.href='qaList.co'">목록으로</button>
-                <button class="writeFaqBtn" onclick="location.href='editQa.co'">수정하기</button>
-                <button class="writeFaqBtn" onclick="location.href='commList.co'">삭제하기</button>
-                <button class="writeFaqBtn" onclick="location.href='replyQa.co'">답글달기</button>
+                <button class="writeFaqBtn" onclick="location.href='qaList.bo'">목록으로</button>
+                <button class="writeFaqBtn" onclick="location.href='editQa.bo'">수정하기</button>
+                <button class="writeFaqBtn" onclick="location.href='commList.bo'">삭제하기</button>
+                <button class="writeFaqBtn" onclick="location.href='replyQa.bo'">답글달기</button>
             </div>
 			</div>
     </section>
