+++ FInalProj/src/main/resources/templates/views/board/qaList.html
@@ -4,7 +4,6 @@
 <meta charset="UTF-8">
 <title>Insert title here</title>
 <link rel="stylesheet" href="css/community.css">
-<link rel="stylesheet" href="resources/css/style.css">
 <script src="http://code.jquery.com/jquery-3.6.1.min.js"></script>
 <script src="https://kit.fontawesome.com/1c67f557d9.js" crossorigin="anonymous"></script>
 </head>
@@ -13,25 +12,31 @@
 <section id="faqSection">
 
         <div>
-            <div id="faqTitle">TALK</div>
+            <div id="faqTitle">1:1문의</div>
         </div>
+        <br><br>
+        <div class="qaInfo">
+        <h4>●문의 전 유의사항●</h4>
+        	<ol>
+            	<li>1. '자주 묻는 질문' 탭에 주로 올라오는 문의들을 따로 정리해놨으니 1:1 문의 전 참고바랍니다.</li>
+            	<li>2. 문의에 대한 답변은 3~7일내 올라온 순서대로 답변해드립니다.</li>
+            	<li>3. 전화문의는 1111-2222 입니다.</li>
+            </ol>
+            </div>
 
         <div class="path">
             <a class="middle" href="#"><img class="homeIconSize" src="image/home.png"></a>
             <img class="sub iconSize" src="image/onenext.png">
             <a class="middle" href="#">커뮤니티</a>
             <img class="sub iconSize" src="image/onenext.png">
-            <a class="middle">TALK</a>
+            <a class="middle">1:1문의</a>
         </div>
 
         <div class="actionArea">
             <div class="writeArea">
-                <button class="writeFaqBtn" onclick="location.href='${ contextPath }/writeComm.co'">글쓰기</button>
+                <button class="writeFaqBtn" onclick="location.href='writeQa.bo'">1:1 질문하기</button>
             </div>
-            <div class="searchArea">
-                <input class="searchInput" type="text" placeholder="검색"/>
-                <a class="searchIconATag" href="#"><img class="searchIcon" src="image/search.png"></a>
-            </div>
+            
         </div>
 
         <table class="faqTable">
@@ -41,80 +46,81 @@
                     <th style="width: 300px;">제목</th>
                     <th>작성자</th>
                     <th>작성일</th>
-                    <th>조회수</th>
                 </tr>
             </thead>
             <tbody>
                 <tr>
                     <td>1</td>
-                    <td>좋은아침입니다~</td>
+                    <td>결제가 안됩니다</td>
                     <td>작성자</td>
                     <td>2023-05-31</td>
-                    <td>0</td>
                 </tr>
                 <tr>
                     <td>2</td>
-                    <td>갈매동 후원소식이에요</td>
+                    <td>오늘 서버쪽에 문제가 있나요?</td>
                     <td>작성자</td>
                     <td>2023-05-31</td>
-                    <td>0</td>
                 </tr>
                 <tr>
                     <td>3</td>
-                    <td>행복하세요 다들~</td>
+                    <td>자원봉사관련 문의요!</td>
                     <td>작성자</td>
                     <td>2023-05-31</td>
-                    <td>0</td>
                 </tr>
                 <tr>
                     <td>4</td>
-                    <td>슬아트공방 다녀왔어요</td>
+                    <td>후원금 환불이 가능할까요?</td>
                     <td>작성자</td>
                     <td>2023-05-31</td>
-                    <td>0</td>
                 </tr>
                 <tr>
                     <td>5</td>
-                    <td>6월6일 사친방지 모임있어요</td>
+                    <td>일반게시판 비속어 신고합니다.</td>
                     <td>작성자</td>
                     <td>2023-05-31</td>
-                    <td>0</td>
                 </tr>
                 <tr>
                     <td>6</td>
-                    <td>이디야 봉화산점 많은이용 바랍니다.</td>
+                    <td>정기후원도 가능한가요?</td>
                     <td>작성자</td>
                     <td>2023-05-31</td>
-                    <td>0</td>
                 </tr>
                 <tr>
                     <td>7</td>
-                    <td>지혜누나는 이제 봄이 올 것 같아요</td>
+                    <td>양로원 봉사모집관련해서 궁금한게 있습니다.</td>
                     <td>작성자</td>
                     <td>2023-05-31</td>
-                    <td>0</td>
                 </tr>
                 <tr>
                     <td>8</td>
-                    <td>은채누나는 베트남에 가 있어요</td>
+                    <td>굿즈판매 안하시나요?</td>
                     <td>작성자</td>
                     <td>2023-05-31</td>
-                    <td>0</td>
                 </tr>
                 <tr>
                     <td>9</td>
-                    <td>도빈이형을 4달만에 보네요!</td>
+                    <td>결제 질문</td>
                     <td>작성자</td>
                     <td>2023-05-31</td>
-                    <td>0</td>
                 </tr>
+                <tr class="reply">
+                	<td>9-1<td>
+          			<div class="replyTitle"><img class="sub iconSize" src="image/reply.png">[Re]결제관련 답변입니다.</div>
+          			<td>관리자</td>
+                    <td>2023-05-31</td>
+        		</tr>
                 <tr>
                     <td>10</td>
-                    <td>지호형, 형주형, 형민이형도 못 뵌지 꽤 됬네요!</td>
+                    <td>신용카드 후원결제가 먹통입니다</td>
                     <td>작성자</td>
                     <td>2023-05-31</td>
-                    <td>0</td>
                 </tr>
+                <tr class="reply">
+                	<td>10-1<td>
+          			<div class="replyTitle"><img class="sub iconSize" src="image/reply.png">[Re]먹통관련 답변입니다.</div>
+          			<td>관리자</td>
+                    <td>2023-05-31</td>
+        		</tr>
             </tbody>
         </table>
