+++ FInalProj/src/main/resources/templates/views/board/replyQa.html
@@ -72,7 +72,7 @@ text-align : center;
 			<input type="password" name="pwd" id="bpwd" class="form-control">
 		</div>
 		<div class="button-container">
-		<input id="subBtn" type="button" value="목록으로" style="float: right;" onclick = "location.href='qaList.co'"/>
+		<input id="subBtn" type="button" value="목록으로" style="float: right;" onclick = "location.href='qaList.bo'"/>
 		<input id="subBtn" type="button" value="답변 등록" style="float: right;" onclick="goWrite(this.form)"/>
 		</div>
 	</form>
