+++ FInalProj/src/main/resources/templates/views/board/writeComm.html
@@ -67,7 +67,7 @@ text-align : center;
 		<br><br> 
 		<textarea id="summernote" name="content"></textarea><br>
 		<div class="button-container">
-		<input id="subBtn" type="button" value="목록으로" style="float: right;" onclick = "location.href='commList.co'"/>
+		<input id="subBtn" type="button" value="목록으로" style="float: right;" onclick = "location.href='commList.bo'"/>
 		<input id="subBtn" type="button" value="글 등록" style="float: right;" onclick="goWrite(this.form)"/>
 		</div>
 	</form>
