+++ FInalProj/src/main/resources/templates/views/board/writeNotice.html
@@ -16,7 +16,7 @@ text-align : center;
 }
 #subBtn {
   /* 생략 */
-  width: 80px;
+  width: 120px;
     height: 30px;
     background-color: #9DC08B;
     color: white;
@@ -41,8 +41,9 @@ text-align : center;
   justify-content: flex-end;
   gap: 10px;
 }
-</style>
 
+
+</style>
 <link rel="stylesheet" href="resources/css/style.css">
 <script src="http://code.jquery.com/jquery-3.6.1.min.js"></script>
 <script src="https://kit.fontawesome.com/1c67f557d9.js" crossorigin="anonymous"></script>
@@ -56,11 +57,12 @@ text-align : center;
 <!-- include summernote-ko-KR -->
 <script src="/resources/js/summernote-ko-KR.js"></script>
 
-<title>문의 작성</title>
+
+<title>공지사항 작성</title>
 </head>
 <body>
 <div th:replace="~{../common/header :: header}"></div>
-<br><br><br><h2 style="text-align: center;" class="title">1:1문의 작성</h2><br><br><br>
+<br><br><br><h2 style="text-align: center;" class="title">공지사항 작성</h2><br><br><br>
 
 <div style="width: 60%; margin: auto;">
 	<form method="post" action="/write">
@@ -68,8 +70,8 @@ text-align : center;
 		<br><br> 
 		<textarea id="summernote" name="content"></textarea><br>
 		<div class="button-container">
-		<input id="subBtn" type="button" value="목록으로" style="float: right;" onclick = "location.href='qaList.co'"/>
-		<input id="subBtn" type="button" value="문의 등록" style="float: right;" onclick="goWrite(this.form)"/>
+		<input id="subBtn" type="button" value="목록으로" style="float: right;" onclick = "location.href='qaList.bo'"/>
+		<input id="subBtn" type="button" value="공지사항 등록" style="float: right;" onclick="goWrite(this.form)"/>
 		</div>
 	</form>
 </div>
@@ -86,5 +88,26 @@ $(document).ready(function() {
 	        lang : 'ko-KR'
 	  });
 	});
+	
+$(document).ready(function() {
+	  // 썸머노트 초기화
+	  $('#summernote').summernote({
+	    placeholder: 'content',
+	    minHeight: 370,
+	    maxHeight: null,
+	    focus: true, 
+	    lang : 'ko-KR'
+	  });
+
+	  // 썸머노트 에디터의 폰트 패밀리 변경
+	  $('.note-editable').css('font-family', 'Arial, sans-serif');
+
+	  // 썸머노트 에디터의 배경색 변경
+	  $('.note-editable').css('background-color', 'lightgray');
+
+	  // 자체 CSS 클래스에 스타일 추가
+	  $('.title').addClass('custom-title');
+	});
+
 </script>
 </html>
