+++ FInalProj/src/main/resources/templates/views/board/writeQa.html
@@ -42,6 +42,7 @@ text-align : center;
   gap: 10px;
 }
 </style>
+
 <link rel="stylesheet" href="resources/css/style.css">
 <script src="http://code.jquery.com/jquery-3.6.1.min.js"></script>
 <script src="https://kit.fontawesome.com/1c67f557d9.js" crossorigin="anonymous"></script>
@@ -55,11 +56,11 @@ text-align : center;
 <!-- include summernote-ko-KR -->
 <script src="/resources/js/summernote-ko-KR.js"></script>
 
-<title>수정</title>
+<title>문의 작성</title>
 </head>
 <body>
 <div th:replace="~{../common/header :: header}"></div>
-<br><br><br><h2 style="text-align: center;" class="title">공지사항 수정하기</h2><br><br><br>
+<br><br><br><h2 style="text-align: center;" class="title">1:1문의 작성</h2><br><br><br>
 
 <div style="width: 60%; margin: auto;">
 	<form method="post" action="/write">
@@ -67,8 +68,8 @@ text-align : center;
 		<br><br> 
 		<textarea id="summernote" name="content"></textarea><br>
 		<div class="button-container">
-		<input id="subBtn" type="button" value="수정취소" style="float: right;" onclick = "location.href='${ contextPath }/noticeList.co'"/>
-		<input id="subBtn" type="button" value="수정하기" style="float: right;" onclick="goWrite(this.form)"/>
+		<input id="subBtn" type="button" value="목록으로" style="float: right;" onclick = "location.href='qaList.bo'"/>
+		<input id="subBtn" type="button" value="문의 등록" style="float: right;" onclick="goWrite(this.form)"/>
 		</div>
 	</form>
 </div>
